\name{print.ptycho}
\alias{print.ptycho}
\title{Print ptycho Object}
\description{Print the sample means in a \code{ptycho} object; do not print the
  attributes.}
\usage{\method{print}{ptycho}(x, ...)}
\arguments{
  \item{x}{Object of class \code{ptycho}}
  \item{\dots}{Additional print arguments}
}
\author{
  Laurel Stell and Chiara Sabatti\cr
  Maintainer: Laurel Stell <lstell@stanford.edu>
}
\seealso{\code{\link{ptycho}}}
